feat(console): migrate Beads to TanStack Query + Suspense (ADR 0013)
Next slice of the console data-layer migration. The Beads surface becomes a
self-contained BeadsPanel on the query layer:

- Reads: useSuspenseQuery(beadsIssuesQuery) — refetches while mounted.
- Mutations: create / start / close / reopen / delete via useMutation,
  invalidating the issues query; delete routes through the App-owned confirm.
- Loading → <Suspense> skeleton, errors → contained <ErrorBoundary> retry card.
- Drops the App-level beads state (beadsIssues/beadsReady/issueDraft/
  beadsBusy/collapsedIssueGroups) + handlers + the vestigial init flow (the
  in-process store is always ready). refreshProjectState now only loads notes.
- Beads helpers extracted to app/beads.ts (shared, no circular import).

Right panel is now Notes (imperative — edit state) · Beads (query) · Goals
(query). E2E: a beads nav test exercises the query-backed list (mock serves a
sample issue). tsc + build clean; 39 E2E green.
