+++ apps/web/src/lib/queries.ts
@@ -7,6 +7,7 @@ import { api } from "./api";
 // stable and hierarchical so a mutation can invalidate a whole subtree.
 export const queryKeys = {
   goals: ["goals"] as const,
+  beadsIssues: ["beads", "issues"] as const,
 };
 
 // Goals the agent works toward (goal mode). Lives in the right sidebar and
@@ -16,5 +17,14 @@ export const goalsQuery = () =>
   queryOptions({
     queryKey: queryKeys.goals,
     queryFn: () => api.goals(),
+    refetchInterval: 5_000,
+  });
+
+// The agent's task board (in-process beads store — always available). Refetches
+// while mounted so the panel tracks issues the agent files/closes mid-turn.
+export const beadsIssuesQuery = () =>
+  queryOptions({
+    queryKey: queryKeys.beadsIssues,
+    queryFn: () => api.beadsIssues(),
     refetchInterval: 5_000,
   });
